Very rough port to support darwin
Right now this is almost entirely using the home-manager
side of things with the darwin bits just being placeholder
and probably broken.

This might have also broken linux, we'll fix that when I pull
these changes down on one of those machines.  I'm almost surprised
that this works and think there's probably better ways of handling
a bunch of stuff, but I'm on mac now with a really functional
fully configured emacs and wezterm and things are working like
a charm!