Add profile for WSL on top of Ubuntu 22.04
This was a very straightforward process with two steps to
setup on top of  my existing WSL setup in order to hack
on stuff when booted into Windows.

1. Run the determinate systems installer.
2. Do a `nix run home-manager/master -- switch --flake .#<profile>`

There were a few steps outside of this in order to
get 1Password SSH Auth working on windows and passed
through to WSL; might need to revisit to see if anything
from those steps can be further operationalized.

Signed-off-by: Paul Osborne <osbpau@gmail.com>